+++ src/main/java/org/mastodon/pool/SingleArrayMemPool.java
@@ -9,14 +9,14 @@ package org.mastodon.pool;
  * @param <A>
  *            the type of the primitive array used in the
  *            {@link MappedElementArray}.
- * 
+ *
  * @author Tobias Pietzsch &lt;tobias.pietzsch@gmail.com&gt;
  */
 public class SingleArrayMemPool< A extends MappedElementArray< A, T >, T extends MappedElement > extends MemPool< T >
 {
 	private final A data;
 
-	public SingleArrayMemPool( final MappedElementArray.Factory< A > arrayFactory, final int capacity, final int bytesPerElement )
+	public SingleArrayMemPool( final MappedElementArray.Factory< A > arrayFactory, final long capacity, final int bytesPerElement )
 	{
 		super( capacity, bytesPerElement );
 		data = arrayFactory.createArray( capacity, this.bytesPerElement );
@@ -24,9 +24,9 @@ public class SingleArrayMemPool< A extends MappedElementArray< A, T >, T extends
 	}
 
 	@Override
-	protected int append()
+	protected long append()
 	{
-		final int index = allocatedSize++;
+		final long index = allocatedSize++;
 		if ( allocatedSize > capacity )
 		{
 			capacity = Math.min( capacity << 1, data.maxSize() );
@@ -44,20 +44,20 @@ public class SingleArrayMemPool< A extends MappedElementArray< A, T >, T extends
 	}
 
 	@Override
-	public void updateAccess( final T access, final int index )
+	public void updateAccess( final T access, final long index )
 	{
 		data.updateAccess( access, index );
 	}
 
 	@Override
-	public void swap( final int index0, final int index1 )
+	public void swap( final long index0, final long index1 )
 	{
 		data.swapElement( index0, data, index1 );
 	}
 
 	/**
 	 * <b>For internal use only!</b>
-	 * 
+	 *
 	 * @return the data array used in this class.
 	 */
 	public A getDataArray()
@@ -69,7 +69,7 @@ public class SingleArrayMemPool< A extends MappedElementArray< A, T >, T extends
 	 * Creates a factory for {@link SingleArrayMemPool}s that use the specified
 	 * {@code arrayFactory} for creating their storage
 	 * {@link MappedElementArray}.
-	 * 
+	 *
 	 * @param arrayFactory
 	 *            the array factory.
 	 * @return a new factory that can create {@link MemPool}.
@@ -86,7 +86,7 @@ public class SingleArrayMemPool< A extends MappedElementArray< A, T >, T extends
 		return new MemPool.Factory< T >()
 		{
 			@Override
-			public MemPool< T > createPool( final int capacity, final int bytesPerElement )
+			public MemPool< T > createPool( final long capacity, final int bytesPerElement )
 			{
 				return new SingleArrayMemPool<>( arrayFactory, capacity, bytesPerElement );
 			}
